Greyed out enemies that lost, added Damage display, Health bars

+++ TheGladiator/Assets/Scripts/Manager/ItemCreationManager.cs
@@ -24,7 +24,7 @@ public class ItemCreationManager : MonoBehaviour {
     public int BaseStats;
 
     protected int HPPoints;
-    protected byte StrPoints;
+    protected short StrPoints;
     protected byte AgiPoints;
     protected byte DexPoints;
     protected short StaPoints;
